requirements.txt
argparse==1.4.0
glob2==0.7
httpx==0.18.2
json5==0.9.5
python-dotenv==0.19.0
regex==2021.8.28